Changed semantics of ht entry() function
In the new semantics entry() will be called even for
the main thread, and an externed variable current_ht_context will
store the current user context running on a hard thread.  For now
current_ht_context only ever points to the main thread at the time
that ht0 is launched.  In the future though, this will also be set
on other hts as they are interrupted for async I/O completions or
signal handling purposes.