Streamlined lithe.c to not copy so much data around in tls's
Adjusted applications to work with new semantics.  Alot easier to reason about
now as well, since 'current_task' is now just a synonym for 'current_uthread'
and there is no notion of 'current_sched'.  The only current scheduler is the
value of 'current_task->sched', and the only book keeping we need to do is make
sure we update this value within the trampoline before jumping to it, as well
as update the trampoline pointer in each task before resuming it.  Still alot
less book keeping than before...