Step one towards removing the trampoline in lithe
So far I've just added the TLS variable next_func to force us to jump to the
trampoline from within vcore_entry rather than from each individual function
that previously did it themselves.  The next step will be to remove the
trampoline completely and just ahve vcore_entry call these functions directly.
Minor issue is that now we will need to keep track of the current scheduler
from within vcore context and not just task context.  Currently the current
scheuler is just a field of the current task, but once we run functions from
directly within vcore_context there will be no current task from which to
figure out which scheduler we need to use.  I plan to figure out a nice way to
handle this in the next commit.