Modify semantics of lithe_sched_enter()/exit() to swap to new task
This required some modifications of the underlying uthread library to make this
cleanly possible.  Essentially, we 'highjack' the current thread context to
start running it as a new uthread (lithe_task).  When we exit, we 'highjack'
the thread context back and start running it again as the original uthread.

One side effect of this method is that the new uthread runs on the same stack
as the original uthread.  Thus no new stack is ever created for this new
uthread.