Introduced new sched functions for task_stack_create()/destroy()
Introducing these cleans up the code for task stack/creation significantly. All
logic for fi/when stacks should be created by the scheduler is now encapsulated
completely within lithe, while the mechanism for doing so is kept in the
scheduler (where it belongs).

Also added a lithe_task_attr_init() function so that atrributes can be
intiialized to default values before setting specific ones for use when
creating a new task.  Not calling this function on before passing a set of
attributes to either lithe_task_create() or lithe_task_init() is an error.

Also added asserts for all callbacks from lithe into the scheduler to make sure
they exist before calling them.