Got rid of callbacks for context_create/destroy()
Instead we now have only a context_init() and context_cleanup(). Similar
changes were made down in the uthread code.  Since the uthread code allocates
and deallocates TLS regions for each uthread, one unfortunate drawback of this
approach is that we now have to make sure any contexts we pass to
context_init() has been pre-zeroed out.  If not, funcky things will happen
since we first check to see if a TLS region already exists, and if so, free the
existing one, then allocate a new one. If the structure is not zeroed out
before hand, we may just try and free garbage, resulting in a segfault.