Removed superfluous main_context var and cleaned up yeilding/blocking
Got rid of the separate 'finished' and 'blocked' flags in the lithe_context
struct and added a 'state' field.  This state field is set to one of the state
values declared at the top of context.h so that the lithe runtime and properly
manage contexts when yielding/blocking/exiting, etc.