Change to creating shared instead of static libs
This change prompted a redesign of how the tls variables get accessed when
changing the tls_desc in the middle of a function.  With O2 optimizations, gcc
doesn't recognize that a tls_desc has been changed, and a cahced tls variable
address may get used incorrectly (only becomes an issue once we introduce PIC
code and gcc tries to optimize out calls to ___tls_get_addr()).