Remove ability to set a vcore tls variable on vcores other than your own
Allowing one to tls variables in another vcore's tls region complicates the
vcore_get/set_tls_var macros once we introduce the notion of migrating threads
across vcores at arbitrary times.  With this change we can encapsulate the
edition of a NO_MIGRATE flag within the macros instead of forcing them to be
outside (since what we really want to protect is the vcoreid previously
passed).