Allow me to have a gh-pages branch under doc folder
What I basically do is clone a copy of the repo at doc/_build/html and then
checkout the gh-pages branch.  This way when I do a 'make html' all of my html
documentation gets dumped into this branch of the repo and I can just cd into
it, add the new files, and commit to make it show up on as my gh-page.

To make sure that a 'git clean -xdf' doesn't blow this away, I check in the
.gitignore and the .nojekyll files from doc/_build/html to make sure this guy
sticks around and the only new content that gets added is from a 'make html'.
I also modify the doc Makefile slightly to make sure that a make clean doesn't
wipe it out either.