Get rid of this uthread highjacking nonsense
I think I had it in there because Ben indicated at one point that the
tls of the current context should be new once a child scheduler was
entered.  As far as I can tell though, this is NOT what we actually
want, so I've removed it.  Now all that happens is that a child becomes
aware of it's new scheduler, but retains everything else.