Redesign of futex code for higher performance
The old futex code used to maintain a single global list for all futexes being
waited on by the app.  We have redesiged this to follow a "list-of-lists"
pattern, where the innermost list only contains futexes all waiting on the
same uaddr.  The outerlist (which could be implemented as a hash), just helps
us find our list of threads waiting on the same futex.  The outerlist is
maintained as a wait free list for higher concurrency, and the innerlist is
protected by a lock, since there usually shouldn't be a whole lot of
contention on it and it simplifies the logic alot.

We also add api calls to optimize for the most commone wake_one() and
wake_all() cases of waking up some futexes.