Change semantics of hart_request() return value
The return value ow indicates how many harts were successfully
requested.  Previously it was supposed to have 'all-or-nothing'
semantics, in that either your full request succeeded (whereby those
harts will trickle in), or it failed (and you will get none of the harts
trickling in).  Under the new semantics, the return value indicates how
many of the requested are guaranteed to pop up in the very near future,
and you can expect to see some (but not necessary all) of the rest of
the requested harts trickle in at some arbitrary point in the future.