Added a parent_data field to the lithe_sched_t
This field is intended for use by the parent of a given scheduler to
allow it to maintain parent-specific state associated with its children.
When a parent receives a callback triggered by its children, all it has
access to is a pointer to that child.  Using this field it can now maintain
parent specific ids for the child, store a pointer back to a more
complicated data structure specific to that child, or do whatever it
wants to really.  Having this field is basically an alternative to
requiring the use of a hash table to redirect us to some piece of data,
given the child pointer.