Change semantics of hart_request to absolute
The previous semantics were incremental, i.e. give me 1 more hart.  the
new semantics are absolute, i.e. the amount I tell you is now the amount
I want, ignore any previous requests.  The old semantics presented
several problems since there was no good way to convey the absolute
number of harts you would actually like to have, and there was no good
way to "cancel" a hart request once it was submitted.
With the new interface, you simply update the value in the requdst with
the total number of harts you'd like to have right now (even if that
number is less than the amount in a previous request).
This doesn't mean you might not temprarily get more (since the interface
is inherently racy), but now your parent has more information on your
exact real-time needs, and can react in response to them.

The new semantics are also strictly more powerful, since an incremental
interface could easily be built on top of it (while doing it the other
way around was not).