Inform the scheduler of a block BEFORE blocking
Previously, the scheduler would be informed of a block AFTER the
call-site specific block function was ran.  This could lead to races
with book-keeping being done in the scheduler.  If a call-site specific
callback was issued, putting the context on a queue, then it was
possible that some other hart could come along and wake up that context
before the scheduler callback for block() was ever issued. This can no
longer happen.