Fix logic for ctx->preferred_vcq
The main context wasn't setting its preferred vcq properly.
Additionally, the check for when to assign it via a call to
get_next_vcq() should check for the existing value being -1, not on the
state being CREATED.  Future Kevin will wonder why just checking CREATED
isn't enough, but it's because libgomp also needs to have it reset to a
new vcq value after being reassociated with a new scheduler and in the
BLOCKED state.  In general, it makes sense to use a sentinal value for
preferred_vcq to decide whether to reset it, rather than relying on the
state variable anyway.