Fix refcounting on root_sched with syscalls
We need to keep separate ref counts for the existence of the scheduler and the
first hart it enters on.  Since we now allow syscalls to block and then
wake a scheduler up again later, its possible to release all harts but
still be alive waiting for syscalls to complete.  Previously we didn't
see this problem because an active scheduler always had at least one
hart, so bundling the schedulers existence and its frist hart into a
single ref count was ok.  Now it is not, hence the separation.

Not separating these counts manifested itself as a bug when waiting for
syscall events to complete. The root scheudler would release all of it
harts, dropping its refcount to 0.  Then when a syscall completion came
in, the handler would poke the scheduler triggering it to put in a hart
request, but that request would never be serviced because the base
shceudler would see a ref count of 0 on the root scheudler, and treat
the hart request as spurious and simply yield it back to the system.
Separating the hart and shceudler refcounts in this patch fixes this
issue.