Need a lock instead of an atomic op
Unfortunately, using an absolute hart request interface forces us to
have a lock here (or modify the hart_request interface to take a
sequence number or something) because two cores might be putting in
requests concurrently.  I was seeing bugs with the underlying request
being put in out-of-order, causing deadlock when a reqeust of '1' got
put in *after* a request for 0, even though they were ordered properly
in the child scheduler.  Using an explicit lock fixes this.