All lithe_context_unblocks() now switch to context sched
Previously we only did this when a syscall completion event came in
(because it was more likely to be in another scheduler when this
happens).  However, normal use of futexes and the like, maay actually
want to synchronize access to things across schedulers.  Graphics magic
does this (for example) with an OMP critical section. When doing things
the old way, I would see contexts pop up in the wrong scheduler
sometimes, messing with their hart count, etc. when those contexts
completed.  It makes sense to do this anyway, and is not harmful.  It is
strictly more powerful.